solocmr/backend/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.1.0
schedule==1.2.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
